+++ beacon_api/utils/data_query.py
@@ -168,7 +168,7 @@ async def fetch_filtered_dataset(db_pool, assembly_id, position, chromosome, ref
                             WHERE a.datasetId=b.datasetId
                             AND b.assemblyId=$3
                             AND {"NOT" if misses else ''} (($8::integer IS NULL OR a.start=$8)
-                            AND ($9 IS NULL OR a.end=$9)
+                            AND ($9::integer IS NULL OR a.end=$9)
                             AND ($10::integer IS NULL OR a.start<=$10) AND ($11::integer IS NULL OR a.start>=$11)
                             AND ($12::integer IS NULL OR a.end>=$12) AND ($13::integer IS NULL OR a.end<=$13)
                             AND coalesce(a.reference LIKE any($7::varchar[]), true)
